#7 add api data to meter reading list

+++ app/src/main/java/io/devmartynov/house/ui/screen/meterReading/components/MeterReadingInfo.kt
@@ -15,7 +15,7 @@ fun MeterReadingInfo(
     value: Float,
     createDate: String,
     isExpired: Boolean,
-    toPayAmount: Int,
+    toPayAmount: Float,
     diffWithPrevValue: Float,
     usageAmount: Float,
 ) {
